Headerコンポーネントを更新し、ナビゲーション項目をNavItemsコンポーネントに移行しました。これにより、コードの可読性が向上し、ナビゲーションの管理が容易になりました。
